Add flexbox layout and padding
In this commit, I followed the instructions in the README file to adjust
the previous state to match the desired outcome using flexbox properties
and padding in the stylesheet, as well as nested div classes within the
html. Admittedly, I got confused so I played with the code in
solution.css to understand why that was the solution compared to what I
previously tried.