Add flexbox layout
In this commit, I followed the instructions in the README file to adjust
the previous state to match the desired outcome using nested div
classes in the html, as well as flexbox properties, padding, and
typography properties in the stylesheet. I did get confused on how to
get all the other elements to sit to the right of the icon, and I did
not consider a gap property to be the obvious solution there, but I feel
better about my flexbox skills having gotten pretty much everything else
before checking the solution.