Add layout
In this commit, I followed the instructions in the README file to adjust
the previous state to match the desired outcome using flexbox and other
properties in the stylesheet. I admittedly got confused in a number of
places, and this will be an exercise that I come back to regularly to
learn from on other projects so that I can get better.