Add layout
In this commit, I followed the instructions in the README file to adjust
the previous state to match the desired outcome using flexbox and other
properties in the stylesheet, and a couple of additional container
classes in the html. Here, I gained a better understanding between the
auto and wrap values within the flex-wrap properties, and I feel more
confident about my flexbox skills overall. Progress!